Add error handling for mycred command and ledger fetch, also added runOnInit for CronJob
