+++ src/handlers/mycred.js
@@ -1,4 +1,5 @@
 const { credEmbed } = require('../embed')
+const { error } = require('../utils')
 
 module.exports = async function mycred(message, pollenData) {
   if (!pollenData) return message.reply('Still preloading pollen files, try again in a minute.')
@@ -13,10 +14,15 @@ module.exports = async function mycred(message, pollenData) {
   const credParticipant = credParticipants.find(p => p.id === accountFound.identity.id)
   const credHistory = credParticipant.credPerInterval
 
-  message.channel.send(`<@${message.author.id}>`)
-  message.channel.send(credEmbed(
-    credParticipant.cred,
-    credHistory[credHistory.length -2],
-    credHistory[credHistory.length - 1]
-  ))
+  try {
+    await message.channel.send(`<@${message.author.id}>`)
+    await message.channel.send(credEmbed(
+      credParticipant.cred,
+      credHistory[credHistory.length -2],
+      credHistory[credHistory.length - 1]
+    ))
+  } catch (err) {
+    error(err)
+    message.reply(`An error has occurred: ${err}`)
+  }
 }
